refact(modularization/subscription-settings): categorizing subscription settings as module
