+++ dashboard/src/modules/settings/subscription/form/fields/update-interval.tsx
