+++ dashboard/src/modules/settings/subscription/subscription-rule.type.ts
